changed the data type for PersonalNumber column in Customers table

+++ CustomerAPI/Entities/Customer.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using System.Numerics;
 using System.Text;
 
 namespace CustomerAPI.Entities
@@ -13,7 +14,7 @@ namespace CustomerAPI.Entities
 
         [Required(ErrorMessage = "Personal Number can't be empty")]
         [MaxLength(12)]
-        public long PersonalNumber { get; set; }
+        public BigInteger PersonalNumber { get; set; }
 
         [Required(ErrorMessage = "Email cannot be empty")]
         [DataType(DataType.EmailAddress, ErrorMessage = "Invalid Email ID")]
